refactor(fleets): change components location
